+++ src/main/kotlin/blackjack/domain/Dealer.kt
@@ -1,0 +1,10 @@
+package blackjack.domain
+
+class Dealer(val deck: Deck) {
+    fun startRound(players: Array<Player>) {
+        for (player in players) {
+            val firstRoundCards = listOf(deck.drawCard(), deck.drawCard())
+            player.hands.addAll(firstRoundCards)
+        }
+    }
+}
