+++ src/main/kotlin/blackjack/domain/Player.kt
@@ -1,0 +1,5 @@
+package blackjack.domain
+
+class Player(val name: String) {
+    val hands = mutableListOf<PokerCard>()
+}
