+++ src/main/kotlin/blackjack/domain/PokerCard.kt
@@ -1,0 +1,3 @@
+package blackjack.domain
+
+class PokerCard(val symbol: PokerSymbol, val value: Int)
